fix(web): rename catalogPagination module to avoid case collision
CatalogPagination.tsx and catalogPagination.ts differed only in their
leading capital, so case-insensitive filesystems collapsed the two and
resolved the component's own import to the wrong module. That failed
'next build' and produced 45 spurious tsc errors across 25 files.

Rename the helpers module to catalogPaginationUtils.ts.
